backend-local/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
requests
numpy
pymupdf
python-multipart
